ml-service/requirements.txt
flask
pandas
mysql-connector-python
scikit-learn
numpy
